fixup refactor(extractor) split off `fixers` to avoid circular import problem
TaskResult has caused a circular import, while it was not usable for
extractors, as they had no access to the TaskResult object.

This change allows tests and unblob to run again.
This is a technical commit to allow for easier review,
should be squashed after review.
